Remove export * as from react-hooks package in favor of explicit exports
