--- packages/react-hooks/src/constants/hooks.ts
@@ -1,7 +1,0 @@
-export const time = {
-  oneSecond: 1 * 1000,
-  oneMinute: 60 * 1000,
-  oneHour: 60 * 60 * 1000
-}
-
-export const ZERO_ADDRESS = '0x0000000000000000000000000000000000000000'
