--- packages/react-hooks/src/constants/index.ts
@@ -1,4 +1,0 @@
-export * from './hooks'
-export * from './queryKeys'
-export * from './swaps'
-export * from './tests'
