--- packages/react-hooks/src/constants/swaps.ts
@@ -1,2 +1,0 @@
-// Address displayed for native tokens by the swap API
-export const NATIVE_TOKEN_ADDRESS_0X_SWAP = '0xeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee'
