--- packages/react-hooks/src/constants/tests.ts
@@ -1,1 +1,0 @@
-export const ACCOUNT_ADDRESS = '0x345458cfD2F0c808455342CD0A6e07a09f893292'
