--- packages/react-hooks/src/hooks/API/index.ts
@@ -1,4 +1,0 @@
-export * from './useGetCoinPrices'
-export * from './useGetCollectiblePrices'
-export * from './useGetExchangeRate'
-export * from './useAPIClient'
