--- packages/react-hooks/src/hooks/Combination/index.ts
@@ -1,2 +1,0 @@
-export * from './useGetSwapPrices'
-export * from './useGetSwapQuote'
