--- packages/react-hooks/src/hooks/Indexer/index.ts
@@ -1,3 +1,0 @@
-export * from './useGetTransactionHistory'
-export * from './useGetTransactionHistorySummary'
-export * from './useIndexerClient'
